Move color tokens out of `FluentUI_common` and into iOS
