+++ Sources/FluentUI_common/Core/Theme/PlatformThemeProviding.swift
@@ -9,7 +9,6 @@ public protocol PlatformThemeProviding {
     /// Returns the platform-appropriate override for a given `ColorToken`.
     /// - Parameters:
     ///   - token: The `ColorToken` whose color should be provided.
-    ///   - defaultColor: The default value for the given `ColorToken`.
-    /// - Returns: The overridden value, or `nil` if no override is needed.
-    static func platformColorValue(_ token: FluentTheme.ColorToken, defaultColor: DynamicColor) -> DynamicColor?
+    /// - Returns: The color value for this token.
+    static func platformColorValue(_ token: FluentTheme.ColorToken) -> DynamicColor
 }
